requirements.txt
Flask==3.1.1
gunicorn==23.0.0
google-cloud-logging==3.12.1
google-cloud-firestore==2.21.0
google-cloud-storage==2.17.0
google-cloud-secret-manager==2.24.0
google-api-python-client==2.178.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-cloud-translate==3.21.1

